ceil -> ceilf double promotion removed

+++ Resampler.h
@@ -88,7 +88,7 @@ class Resampler {
                 float dist=successorIndex-_cPos;
                     
                 float distScaled=dist*_overSamplingFactor;
-                int32_t rightIndex=abs((int32_t)(ceil(distScaled))-_overSamplingFactor*_halfFilterLength);   
+                int32_t rightIndex=abs((int32_t)(ceilf(distScaled))-_overSamplingFactor*_halfFilterLength);   
                 const int32_t indexData=successorIndex-_halfFilterLength;
                 if (indexData>=0){
                     for (uint8_t i =0; i< NOCHANNELS; i++){
@@ -108,11 +108,11 @@ class Resampler {
                         *(si1Ptr++)=*ip[i]++**fPtr;
                     }
                     fPtr-=_overSamplingFactor;          
-                    rightIndex=(int32_t)(ceil(distScaled))+_overSamplingFactor;     //needed below  
+                    rightIndex=(int32_t)(ceilf(distScaled))+_overSamplingFactor;     //needed below  
                 }
                 else {
                     memset(si1, 0, NOCHANNELS*sizeof(float));
-                    rightIndex=(int32_t)(ceil(distScaled));     //needed below
+                    rightIndex=(int32_t)(ceilf(distScaled));     //needed below
                 }
                 for (uint16_t i =0 ; i<_halfFilterLength; i++){
                     if(ip[0]==_endOfBuffer[0]){
@@ -147,8 +147,8 @@ class Resampler {
                     }
                     fPtr+=_overSamplingFactor;
                 }
-                const float w0=ceil(distScaled)-distScaled;
-                const float w1=1.-w0;
+                const float w0=ceilf(distScaled)-distScaled;
+                const float w1=1.f-w0;
                 si0Ptr=si0;
                 si1Ptr=si1;
                 for (uint8_t i =0; i< NOCHANNELS; i++){
